+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 
 * Implemented [end-point](https://www.textcontrol.com/blog/2018/10/01/new-reportingcloud-endpoint-appending-documents/):
     * `/v1/document/append`
-* Added demos `append-document-basic.php` and `append-document-advanced.php` to exemplify `appendDocument($documentsData, $returnFormat, $documentSettings)`.    
+* Added `/demo/append-document-basic.php` and `/demo/append-document-advanced.php` to exemplify `appendDocument($documentsData, $returnFormat, $documentSettings)`.
+* Switched from username and password authentication to API key authentication in all demos and tests. The former authentication is deprecated and will be removed in a future version.    
 
 ## 1.8.0 - 2018-08-02